Rewrite NPAPI plugin for Safari using chrome plugin backend
