--- English.lproj/InfoPlist.strings
@@ -1,3 +1,0 @@
-﻿/* Localized versions of Info.plist keys */
-
-NSHumanReadableCopyright = "© sk, 2010";
