+++ Makefile.Darwin
@@ -3,20 +3,18 @@ include VERSION.mk
 
 SIGNER ?= Riigi Infosüsteemi Amet
 
-PROJ = xcodebuild -project esteid-plugin.xcodeproj VERSION=$(VERSION) BUILD_NUMBER=$(BUILD_NUMBER)
+PROJ = xcodebuild -project browser-token-signing.xcodeproj VERSION=$(VERSION) BUILD_NUMBER=$(BUILD_NUMBER)
 
 TMPROOT = $(PWD)/tmp
 TARGET = $(TMPROOT)/Library/Internet\ Plug-Ins/esteidfirefoxplugin.bundle
 PKG = firefox-token-signing.pkg
 
-OPENSSL ?= $(PWD)/target
-
 build: default
 
 default: pkg
 
-$(TARGET): $(OPENSSL)
-	$(PROJ) DSTROOT=$(TMPROOT) OPENSSL=$(OPENSSL) install
+$(TARGET):
+	$(PROJ) DSTROOT=$(TMPROOT) install
 
 clean:
 	rm -rf $(TMPROOT)
@@ -41,12 +39,3 @@ signed: codesign
 		--install-location / \
 		--sign "Developer ID Installer: $(SIGNER)" \
 		$(PKG)
-
-# FIXME: needs universal
-$(OPENSSL):
-	test -e openssl || git clone --depth=1 https://github.com/openssl/openssl.git -b OpenSSL_1_0_2-stable
-	(cd openssl \
-	&& KERNEL_BITS=64 ./config --prefix=$(OPENSSL) -mmacosx-version-min=10.9 no-shared no-ssl2 no-idea no-dtls no-psk no-srp no-apps \
-	&& make depend \
-	&& make \
-	&& make install_sw)
