+++ common/BinaryUtils.cpp
