+++ common/BinaryUtils.h
