+++ ie/ie-token-signing.vcxproj
@@ -236,14 +236,14 @@
     <ClCompile Include="EstEIDCertificate.cpp" />
     <ClCompile Include="EstEIDIEPluginBHO.cpp" />
     <ClCompile Include="esteidpluginie_i.c" />
-    <ClCompile Include="BinaryUtils.cpp" />
-    <ClCompile Include="Labels.cpp" />
-    <ClCompile Include="Logger.cpp" />
+    <ClCompile Include="..\common\BinaryUtils.cpp" />
+    <ClCompile Include="..\common\Labels.cpp" />
+    <ClCompile Include="..\common\Logger.cpp" />
     <ClCompile Include="NativeCertificateSelector.cpp" />
     <ClCompile Include="NativeSigner.cpp" />
     <ClCompile Include="PinDialog.cpp" />
     <ClCompile Include="PKCS11CertificateSelector.cpp" />
-    <ClCompile Include="PKCS11Path.cpp" />
+    <ClCompile Include="..\common\PKCS11Path.cpp" />
     <ClCompile Include="PKCS11Signer.cpp" />
     <ClCompile Include="Signer.cpp" />
   </ItemGroup>
@@ -264,15 +264,15 @@
     <ClInclude Include="EstEIDCertificate.h" />
     <ClInclude Include="EstEIDIEPluginBHO.h" />
     <ClInclude Include="HostExceptions.h" />
-    <ClInclude Include="Labels.h" />
-    <ClInclude Include="Logger.h" />
+    <ClInclude Include="..\common\Labels.h" />
+    <ClInclude Include="..\common\Logger.h" />
     <ClInclude Include="NativeCertificateSelector.h" />
     <ClInclude Include="NativeSigner.h" />
     <ClInclude Include="PinDialog.h" />
-    <ClInclude Include="PKCS11CardManager.h" />
+    <ClInclude Include="..\common\PKCS11CardManager.h" />
     <ClInclude Include="PKCS11CertificateSelector.h" />
-    <ClInclude Include="PKCS11Path.h" />
-    <ClInclude Include="PKCS11Signer.h" />
+    <ClInclude Include="..\common\PKCS11Path.h" />
+    <ClInclude Include="..\common\PKCS11Signer.h" />
     <ClInclude Include="Resource.h" />
     <ClInclude Include="Signer.h" />
     <ClInclude Include="stdafx.h" />
