+++ npapi/Info.plist
